==> 12a Creating a csv without Ordinal Encoding.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

pd.set_option('display.max_columns', None)

df = pd.read_csv('mandar_gurgaon_properties_missing_value_imputation.csv')

df.shape

df.head()

train_df = df.drop(columns=['society','price_per_sqft'])

train_df.head()

X_label = train_df.drop('price', axis=1)
y_label = train_df['price']

X_label

y_label

export_df = X_label.drop(columns=['pooja room', 'study room', 'others'])
export_df['price'] = y_label

export_df

# we will convert the luxury score into categorical values - Low, Medium, High

def categorize_luxury(score):
    if 0 <= score < 50:
        return "Low"
    elif 50 <= score < 150:
        return "Medium"
    elif 150 <= score <= 175:
        return "High"
    else:
        return None  # or "Undefined" or any other label for scores outside the defined bins

export_df['luxury_category'] = export_df['luxury_score'].apply(categorize_luxury)

# converting the floors into categories

def categorize_floor(floor):
    if 0 <= floor <= 2:
        return "Low Floor"
    elif 3 <= floor <= 10:
        return "Mid Floor"
    elif 11 <= floor <= 51:
        return "High Floor"
    else:
        return None  # or "Undefined" or any other label for floors outside the defined bins

export_df['floor_category'] = export_df['floorNum'].apply(categorize_floor)

export_df

export_df.drop(columns=['floorNum','luxury_score'],inplace=True)

export_df

export_df.to_csv('mandar_gurgaon_properties_post_feature_selection_v2.csv', index=False)

df1 = pd.read_csv('mandar_gurgaon_properties_post_feature_selection_v2.csv')
df2 = pd.read_csv('gurgaon_properties_post_feature_selection_v2.csv')

import pandas as pd

# Assuming df1 and df2 are your DataFrames

# Perform an inner join to find common rows
common_rows = pd.merge(df1, df2, how='inner')

# Count the number of common rows
number_of_common_rows = len(common_rows)

print("Number of identical rows in both dataframes:", number_of_common_rows)


df1

df2
