==> 13_baseline model.ipynb <==
import numpy as np
import pandas as pd

df = pd.read_csv('mandar_gurgaon_properties_post_feature_selection.csv')

df.head()

### we are going to apply linear regression for this dataset
 since we are using a linear model,
 so we will use one hot encoding for categorical column

 sector has 104 categories - so from sector, we will have 104 new columns

 one hot encode -> sector, balcony, agePossession, furnishing type, luxury category, floor category

 ie 3 tasks to be done before applying linear regression
 1. One hot encoding
 2. Standard Scaling
 3. Price column is right skewed. So we will use log-transformation on it so that it will be somewhat normally distributed


X = df.drop(columns=['price'])
y = df['price']

from sklearn.model_selection import KFold, cross_val_score
from sklearn.linear_model import LinearRegression
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import OneHotEncoder, StandardScaler
from sklearn.compose import ColumnTransformer
from sklearn.svm import SVR

columns_to_encode = ['sector', 'balcony', 'agePossession', 'furnishing_type', 'luxury_category', 'floor_category']

# Applying the log1p transformation to the target variable
y_transformed = np.log1p(y)

# Creating a column transformer for preprocessing
preprocessor = ColumnTransformer(
    transformers=[
        ('num', StandardScaler(), ['property_type', 'bedRoom', 'bathroom', 'built_up_area', 'servant room', 'store room']),
        ('cat', OneHotEncoder(drop='first', handle_unknown='ignore'), columns_to_encode)
    ], 
    remainder='passthrough'
)

# Creating a pipeline

# ie first all the preprocessing is done
# and then Linear regression is done
pipeline = Pipeline([
    ('preprocessor', preprocessor),
    # ('regressor', LinearRegression())
    
    ('regressor', SVR(kernel='rbf'))
])

# K-fold cross-validation
kfold = KFold(n_splits=10, shuffle=True, random_state=42)
scores = cross_val_score(pipeline, X, y_transformed, cv=kfold, scoring='r2')

scores.mean()
# this means hamara Linear Regression se starting point hai 85% ie baseline model ki accuracy 85% ie yaha se hum isko aur badhayenge
# this is 88% accuracy for SVM

scores.std()

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X,y_transformed,test_size=0.2,random_state=42)

pipeline.fit(X_train,y_train)

y_pred = pipeline.predict(X_test)

y_pred = np.expm1(y_pred)
# expm1 means we had transformed the output column as log-transformation, so to get the original value back, we used expm1

from sklearn.metrics import mean_absolute_error
mean_absolute_error(np.expm1(y_test),y_pred)

## The mean absolute error for Linear Regression is 0.62 ie on average our model is making mistake of 62 lakhs. ie if a 3bhk flat is 1 crore, then there is a chance that our model will say that the flat is 1.62 crore or 0.46 crore. Therefore we need to reduce this.

## The mean absolute error for SVM is 0.52 ie it is reduced just by changing the model from linear regression to SVM. Therefore we need to reduce this more.
