==> 17_insights-module.ipynb <==
import numpy as np
import pandas as pd

from sklearn.model_selection import KFold, cross_val_score
from sklearn.linear_model import LinearRegression,Ridge
from sklearn.pipeline import Pipeline
from sklearn.preprocessing import OneHotEncoder, StandardScaler, OrdinalEncoder
from sklearn.compose import ColumnTransformer

from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_absolute_error

from sklearn.decomposition import PCA

### In insights module, we want to explain the user that which features affect the price and by what factor do they affect it

#### ie we have to train a model again for doing inference. Some models are not good at inference like Deep learning models. While some are good for inference like Linear Regression

#### Therefore, for our insights module, we will use Linear Regression for inference

df = pd.read_csv('mandar_gurgaon_properties_post_feature_selection_v2.csv').drop(columns=['store room','floor_category','balcony'])

# we dropped these 3 columns 'store room','floor_category','balcony' because the r2 score was not changing significantly due to these columns

# new_df = pd.read_csv('mandar_gurgaon_properties_post_feature_selection_v2.csv')
# new_df

# new_df['sector'] = new_df['sector'].str.replace('sector 17a','sector 17')
# new_df['sector'] = new_df['sector'].str.replace('sector 17b','sector 17')
# new_df['sector'] = new_df['sector'].str.replace('sector 37c','sector 37')
# new_df['sector'] = new_df['sector'].str.replace('sector 88b','sector 88')
# new_df['sector'] = new_df['sector'].str.replace('sector 99a','sector 99')
# new_df['sector'] = new_df['sector'].str.replace('sector 10a','sector 10')

# new_df.to_csv('mandar_gurgaon_properties_post_feature_selection_v2.csv',index=False)

df.head()

# df['sector'].unique().tolist()

# df['sector'].value_counts()

#### We will train a Linear Regression model on this data. ie our goal is to get the best r2 score, because we want to understand that how much variance our features are explaining about the target column

# df = pd.read_csv('mandar_gurgaon_properties_post_feature_selection_v2.csv')

# df['sector'] = df['sector'].str.replace('sector 36a','sector 36')
# df['sector'] = df['sector'].str.replace('sohna road road','sohna road')
# df['sector'] = df['sector'].str.replace('sector 3 phase 2','sector 3')
# df['sector'] = df['sector'].str.replace('sector 3 phase 3 extension','sector 3')


# df.to_csv('mandar_gurgaon_properties_post_feature_selection_v2.csv',index=False)


# 0 -> unfurnished
# 1 -> semifurnished
# 2 -> furnished

# Numerical = bedRoom, bathroom, built_up_area, servant room   --> here we dont have to apply transformations
# Ordinal = property_type, furnishing_type, luxury_category    --> ordinal encoding is used where is natural order like low<medium<high
# OHE = sector, agePossession                                  --> here there is no natural order, so apply OHE

# to simplify, we are converting these 5 agePossession categories into 3 categories

df['agePossession'] = df['agePossession'].replace(
    {
        'Relatively New':'new',
        'Moderately Old':'old',
        'New Property' : 'new',
        'Old Property' : 'old',
        'Under Construction' : 'under construction'
    }
)

df.head()

# doing ordinal encoding for property_type

df['property_type'] = df['property_type'].replace({'flat':0,'house':1})

df.head()

df['luxury_category'] = df['luxury_category'].replace({'Low':0,'Medium':1,'High':2})

df.head()

# this is doing OHE on sector and agePossession columns

new_df = pd.get_dummies(df,columns=['sector','agePossession'],drop_first=True)

new_df

X = new_df.drop(columns=['price'])
y = new_df['price']

y_log = np.log1p(y)
# performing log transform on output column as it was becoming nearly normal distribution

y_log

scaler = StandardScaler()

X_scaled = scaler.fit_transform(X)

X_scaled = pd.DataFrame(X_scaled,columns=X.columns)

X_scaled

kfold = KFold(n_splits=10, shuffle=True, random_state=42)
scores = cross_val_score(LinearRegression(), X_scaled, y_log, cv=kfold, scoring='r2')

scores.mean(),scores.std()

lr = LinearRegression()
ridge = Ridge(alpha=0.0001)

lr.fit(X_scaled,y_log)

ridge.fit(X_scaled,y_log)

lr.coef_.shape

coef_df = pd.DataFrame(ridge.coef_.reshape(1,113),columns=X.columns).stack().reset_index().drop(columns=['level_0']).rename(columns={'level_1':'feature',0:'coef'})

coef_df

### We have got the weights of the coeficients after training. Below is the regression analysis

# 1. Import necessary libraries
import statsmodels.api as sm

# 2. Add a constant to X
X_with_const = sm.add_constant(X_scaled)

# 3. Fit the model
model = sm.OLS(y_log, X_with_const).fit()

# 4. Obtain summary statistics
print(model.summary())

y_log.std()

X_scaled['bedRoom'].std()

0.21 * (0.557/1)

np.expm1(0.030)

2.4726962617564903e-05 * 100
